requirements.txt
Django>=5.2,<5.3
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
gunicorn>=22.0,<23.0
uvicorn>=0.30,<0.31
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
